s/or/and/ in "greater of X or Y" expressions
...as well as "lesser of X or Y" expressions.

(apologies for the pedantry)

The spec has a few instances of "...the greater of X or Y"

Strictly speaking, that should say "and", not "or", I think.

Mathematically speaking:
* "X **and** Y" represents a set of two values which we can perform a "give-me-the-greatest-value" operation on.
*...whereas "X **or** Y" represents a single (indeterminate) value, and it doesn't really make sense to perform a give-me-the-greatest-value operation on that single indeterminate value.

Also, consistency-wise (and as a sanity-check on my instinct here): CSS specs do seem to say "the greatest of X **and** Y" (rather than "or"), when they use this sort of phrasing.

 e.g.:
> the greater of the number determined by the table-column elements **and** the number determined by the first row,

https://www.w3.org/TR/2011/PR-CSS2-20110412/tables.html

> the greater of the resolved min-width, CAPMIN, and GRIDMIN.

https://www.w3.org/TR/css-tables-3/

> the greater of the two applicable Membership fees (for "A" and "B").

https://www.w3.org/Consortium/membership-faq