Adding mousemove and pointermove after mousedown and pointerdown to be excluding inputs
Right now, the mousemove and pointermove events are not excluding inputs, but we want to set the mousemove and pointerdown events to be excluding inputs when they trigger dragging element around or resizing elements but not scrolling. These mousemove or pointermove events fired after mousedown or pointerdown event, but without pointercancel event will make dragging or resizing happen.